+++ src/common/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "common"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+edition.workspace = true
+license.workspace = true
+exclude.workspace = true
+
+[dependencies]
+bytes = { workspace = true }
+mlua = { workspace = true }
+serde_json = { workspace = true }
