+++ src/common/README.md
@@ -1,0 +1,3 @@
+# Common
+
+This is an intermediate crate used between util and the main binary.
