+++ src/main/src/requests.rs
@@ -1,9 +1,9 @@
-use crate::common::BodyLua;
 use axum::{
     body::Body,
     extract::{FromRequest, Multipart, State},
     http::{request::Parts, Request},
 };
+use common::BodyLua;
 use mlua::UserData;
 use std::collections::HashMap;
 use tokio::io::AsyncWriteExt;
