+++ src/main/src/startup.rs
@@ -1,5 +1,3 @@
-use mlua::LuaSerdeExt;
-
 pub static LUA: std::sync::LazyLock<mlua::Lua> = std::sync::LazyLock::new(mlua::Lua::new);
 
 use clap::{command, crate_authors, crate_name, crate_version, Parser};
@@ -31,7 +29,7 @@ enum AstraCLI {
 
 pub async fn init() {
     let lua = &LUA;
-    let lib = include_str!("./lua/astra_bundle.lua");
+    let lib = include_str!("../../lua/astra_bundle.lua");
 
     #[allow(clippy::expect_used)]
     lua.load(lib)
@@ -39,7 +37,7 @@ pub async fn init() {
         .await
         .expect("Couldn't add prelude");
 
-    if let Err(e) = crate::utils::register_utils(lua).await {
+    if let Err(e) = utils::register_utils(lua).await {
         println!("Error setting the util functions: {e}");
     }
 
@@ -134,75 +132,4 @@ async fn register_run_function() {
             println!("Could not insert the function for astra_internal__start_server: {e}");
         }
     }
-}
-
-#[derive(Debug, Clone)]
-pub struct BodyLua {
-    #[allow(unused)]
-    pub body: bytes::Bytes,
-    pub body_string: String,
-}
-impl BodyLua {
-    pub fn new(bytes: bytes::Bytes) -> Self {
-        let body_string = String::from_utf8_lossy(&bytes).to_string();
-
-        Self {
-            body: bytes,
-            body_string,
-        }
-    }
-}
-impl mlua::UserData for BodyLua {
-    fn add_methods<M: mlua::UserDataMethods<Self>>(methods: &mut M) {
-        methods.add_method("text", |_, this, ()| Ok(this.body_string.clone()));
-
-        methods.add_method("json", |_, this, ()| {
-            match serde_json::from_str::<serde_json::Value>(&this.body_string) {
-                Ok(body_json) => Ok(LUA.to_value(&body_json)?),
-                Err(e) => Err(mlua::Error::runtime(format!(
-                    "Could not parse the body as JSON: {e:#?}"
-                ))),
-            }
-        });
-    }
-}
-
-#[allow(unused)]
-pub mod macros {
-    macro_rules! impl_deref {
-        ($struct:ty,$type:ty) => {
-            impl std::ops::Deref for $struct {
-                type Target = $type;
-
-                fn deref(&self) -> &Self::Target {
-                    &self.0
-                }
-            }
-            impl std::ops::DerefMut for $struct {
-                fn deref_mut(&mut self) -> &mut Self::Target {
-                    &mut self.0
-                }
-            }
-        };
-    }
-
-    macro_rules! impl_deref_field {
-        ($struct:ty,$type:ty,$field:ident) => {
-            impl std::ops::Deref for $struct {
-                type Target = $type;
-
-                fn deref(&self) -> &Self::Target {
-                    &self.$field
-                }
-            }
-            impl std::ops::DerefMut for $struct {
-                fn deref_mut(&mut self) -> &mut Self::Target {
-                    &mut self.$field
-                }
-            }
-        };
-    }
-
-    pub(crate) use impl_deref;
-    pub(crate) use impl_deref_field;
 }
