+++ src/utils/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "utils"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+edition.workspace = true
+license.workspace = true
+exclude.workspace = true
+
+[dependencies]
+mlua = { workspace = true }
+axum = { workspace = true }
+tokio = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+bytes = { workspace = true }
+chrono = { version = "0.4.39", features = ["serde"] }
+reqwest = { version = "0.12.11", features = [
+    "json",
+    "rustls-tls",
+    "charset",
+    "http2",
+    "macos-system-configuration",
+], default-features = false }
+sqlx = { version = "0.8.2", features = [
+    "runtime-tokio",
+    "tls-rustls",
+    "postgres",
+    "json",
+    "chrono",
+    "uuid",
+] }
+toml = "0.8.19"
+uuid = { version = "1.11.0", features = ["serde", "v4"] }
+common = { path = "../common" }
