+++ src/utils/src/lib.rs
@@ -4,7 +4,7 @@ mod http_client_request;
 mod tasks;
 
 pub trait LuaUtils {
-    async fn register_to_lua(lua: &mlua::Lua) -> mlua::Result<()>;
+    fn register_to_lua(lua: &mlua::Lua) -> impl std::future::Future<Output = mlua::Result<()>>;
 }
 
 pub async fn register_utils(lua: &mlua::Lua) -> mlua::Result<()> {
